requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
pg8000
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
